Split session.py: primitive in session.py, orchestration in session_manager.py
session.py (160 lines): Session protocol, LLMSession, NullSession — just
the conversation primitive.

session_manager.py (250 lines): SessionEntry, SessionContext, SessionOutcome,
SessionStrategy, SessionManager, resolve_session_strategy, make_session_lm —
the orchestration layer.

A reader opening session.py sees "Session = send/fork/reset" and nothing else.
